requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
google-api-core==2.14.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
llama-index==0.9.8.post1
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.3.5
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pgvector==0.2.4
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.23.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.3
